+to cleanup
I've used a big test suite declared as a separate core and run many of 
the arms against it. Instead of many individual tests a programmatically 
run through the whole test suite. It makes the code shorter and lets you 
focus on the test that are used in one place, but makes the logic of 
handling parsing the test suite and comparing agains expected values a 
bit more complicated.